+++ seed/static/seed/js/controllers/inventory_detail_analyses_modal_controller.js
@@ -24,7 +24,12 @@ angular.module('BE.seed.controller.inventory_detail_analyses_modal', [])
       analyses_service,
       inventory_ids,
     ) {
-      // $scope.meters = meters;
+      $scope.bsyncr_models = [
+        {model_type: 'Simple Linear Regression'},
+        {model_type: 'Three Parameter Linear Model Cooling'},
+        {model_type: 'Three Parameter Linear Model Heating'},
+        {model_type: 'Four Parameter Linear Model'}
+      ];
 
       function initialize_new_analysis () {
         $scope.new_analysis = {name: null, service: null, configuration: null};