+++ seed/static/seed/js/services/inventory_service.js
@@ -20,7 +20,7 @@ angular.module('BE.seed.service.inventory', []).factory('inventory_service', [
       total_taxlots_for_user: 0
     };
 
-    inventory_service.get_properties = function (page, per_page, cycle, profile_id, property_view_ids) {
+    inventory_service.get_properties = function (page, per_page, cycle, profile_id, property_view_ids, save_last_cycle=true) {
 
       var params = {
         organization_id: user_service.get_organization().id,
@@ -34,7 +34,9 @@ angular.module('BE.seed.service.inventory', []).factory('inventory_service', [
         var lastCycleId = inventory_service.get_last_cycle();
         if (_.has(cycle, 'id')) {
           params.cycle = cycle.id;
-          inventory_service.save_last_cycle(cycle.id);
+          if (save_last_cycle === true) {
+            inventory_service.save_last_cycle(cycle.id);
+          }
         } else if (_.includes(validCycleIds, lastCycleId)) {
           params.cycle = lastCycleId;
         }
@@ -249,7 +251,7 @@ angular.module('BE.seed.service.inventory', []).factory('inventory_service', [
     };
 
 
-    inventory_service.get_taxlots = function (page, per_page, cycle, profile_id, inventory_ids) {
+    inventory_service.get_taxlots = function (page, per_page, cycle, profile_id, inventory_ids, save_last_cycle=true) {
       var params = {
         organization_id: user_service.get_organization().id,
         page: page,
@@ -262,7 +264,9 @@ angular.module('BE.seed.service.inventory', []).factory('inventory_service', [
         var lastCycleId = inventory_service.get_last_cycle();
         if (cycle) {
           params.cycle = cycle.id;
-          inventory_service.save_last_cycle(cycle.id);
+          if (save_last_cycle === true) {
+            inventory_service.save_last_cycle(cycle.id);
+          }
         } else if (_.includes(validCycleIds, lastCycleId)) {
           params.cycle = lastCycleId;
         }
@@ -498,6 +502,37 @@ angular.module('BE.seed.service.inventory', []).factory('inventory_service', [
       localStorage.setItem('detailProfiles.' + key, JSON.stringify(profiles));
     };
 
+    inventory_service.get_property_column_names_for_org = function (org_id) {
+      return $http.get('/api/v3/columns/', {
+        params: {
+          inventory_type: 'property',
+          organization_id: org_id,
+          only_used: false,
+          display_units: true
+        }
+      }).then(function (response) {
+        let property_columns = response.data.columns.filter(column => column.table_name == 'PropertyState');
+        return property_columns.map(a => {
+          return { 'column_name': a.column_name, 'display_name': a.display_name };
+        });
+      });
+    };
+
+    inventory_service.get_taxlot_column_names_for_org = function (org_id) {
+      return $http.get('/api/v3/columns/', {
+        params: {
+          inventory_type: 'taxlot',
+          organization_id: org_id,
+          only_used: false,
+          display_units: true
+        }
+      }).then(function (response) {
+        let taxlot_columns = response.data.columns.filter(column => column.table_name == 'TaxLotState');
+        return taxlot_columns.map(a => {
+          return { 'column_name': a.column_name, 'display_name': taxlot_columns.find(x => x.column_name == a.column_name).display_name };
+        });
+      });
+    };
 
     inventory_service.get_property_columns = function () {
       return inventory_service.get_property_columns_for_org(user_service.get_organization().id);