+++ seed/views/v3/analyses.py
@@ -95,12 +95,12 @@ class AnalysisViewSet(viewsets.ViewSet):
             analyses_queryset = (
                 Analysis.objects.filter(organization=organization_id, analysispropertyview__property=property_id)
                 .distinct()
-                .order_by('id')
+                .order_by('-id')
             )
         else:
             analyses_queryset = (
                 Analysis.objects.filter(organization=organization_id)
-                .order_by('id')
+                .order_by('-id')
             )
         for analysis in analyses_queryset:
             property_view_info = analysis.get_property_view_info(property_id)
@@ -176,6 +176,26 @@ class AnalysisViewSet(viewsets.ViewSet):
             analysis = Analysis.objects.get(id=pk, organization_id=organization_id)
             pipeline = AnalysisPipeline.factory(analysis)
             pipeline.stop()
+            return JsonResponse({
+                'status': 'success',
+            })
+        except Analysis.DoesNotExist:
+            return JsonResponse({
+                'status': 'error',
+                'message': 'Requested analysis doesn\'t exist in this organization.'
+            }, status=HTTP_409_CONFLICT)
+
+    @swagger_auto_schema(manual_parameters=[AutoSchemaHelper.query_org_id_field()])
+    @require_organization_id_class
+    @api_endpoint_class
+    @ajax_request_class
+    @has_perm_class('requires_member')
+    def destroy(self, request, pk):
+        organization_id = int(request.query_params.get('organization_id', 0))
+        try:
+            analysis = Analysis.objects.get(id=pk, organization_id=organization_id)
+            pipeline = AnalysisPipeline.factory(analysis)
+            pipeline.delete()
             return JsonResponse({
                 'status': 'success',
             })